MB-30040: Use default layout instead of system for Boost
Patch a86ffbf changed the install layout of Boost on Windows from
'versioned' to 'system'; to match the layout used by *ix so we'd have
the same layout for both OSs:

    system layout:
        include/boost/

    versioned layout:
        include/boost-1.67/boost/

However, CMake's FindBoost module (as used by folly) assumes that the
default layout is used - i.e. 'versioned' on Windows - and hence fails
to correctly link targets using boost when system is chosen.

As such, we need to revert back to the default layout to keep CMake
happy; this bumps the version to 1.67-cb3.

As a consequence this requires that Couchbase Server's search or Boost
is updated, which will occur in a subsequent patch when we move over
to 1.67-cb3.

Change-Id: I5a0304820691e284a67b31699d5a522abf111e48
Reviewed-on: http://review.couchbase.org/101234
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>