MB-30040: Use standard FindBoost to locate Boost; move to 1.67-cb3
With the introduction of Folly we need some of the compiled Boost
libraries (and not just the header-only libraries).

Folly uses the standard CMake 'FindBoost' module; which expects the
default boost installation layout on each platform - 'system' on *ix;
and 'versioned' on Windows - e.g.

   system layout:
        include/boost/
        lib/libboost_atomic.lib

    versioned layout:
        include/boost-1.67/boost/
        lib/libboost_atomic-vc140-mt-x32-1_67.lib

This differs from what we historically used; which was system layout
for all platforms.

Cbdep boost-1.67-cb3 switches to the default layout (system:*ix,
versioned:Windows) for Folly; therefore we need to search in different
locations per-platform and per-Boost version in CB Server's build
also.

Therefore move to 1.67-cb3 and change our FindCouchbaseBoost module to
simply call down to FindBoost for all the heavy work; and just
maintain the Couchbase-specific logic in FindCouchbaseBoost (checks
for supported platform; and backwards compatibilty BOOST_INCLUDE_DIR
variable.

Change-Id: Idaebe84bddb1662fc205d7c569f940ba5dea8dae
Reviewed-on: http://review.couchbase.org/101236
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>