MB-32016: Install sanitizer runtime libraries to <prefix>/lib
When building with the sanitizers; they add a runtime dependancy of
libasan / libubsan to any enabled binary target. These sanitizer
libraries normally reside in the toolchain's search path
(e.g. /usr/local/lib64/libasan.so) which is fine as long as you're
running the binary on the machine it was built onq.

However, when we want to package up (e.g. RPM) and install on a
different machine, the sanitizer runtimes will not be present unless
the toolchain happens to be installed on that machine. Essentially
this is the same situation we have with libstdc++ - we need to install
libasan / libubsan etc into /opt/couchbase/lib so binaries can find
them without the complete compiler toolchain.

As such, if ASAN / UBSAN are enabled; add a install() target for each
library which is enabled.

Change-Id: I7a6158553e53f719c7e8651f05bab98c1a9d1c19
Reviewed-on: http://review.couchbase.org/101849
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>