Extend REBAR() to support "rebar eunit".
A special case of "rebar eunit" that only compiles a test version of
the code but doesn't actually run the code. This allows the test code
to pick the parameters that are used by the tested code at
compile-time.

Additional benefit is slightly faster compile times: normal "make all
install" can be made to skip compilation of tests themselves and their
dependencies altogether.

Change-Id: I75aa8a13d9e28c1912adde75ad7fad8e66b5adc9
Reviewed-on: http://review.couchbase.org/103926
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>