MB-32993: Install PyInstaller stuff into lib/python
PyInstaller seems to modify some libraries when copying them into the
single-directory package, especially on MacOS. As other parts of the
system also depend on those libraries, we now isolate the PyInstaller
versions into a separate directory. This results in about 100MB on-disk
size increate due to library duplication, but is required for safety.

Change-Id: I8e0ee2d65400e82ad105a3af0f388bfcfda3b082
Reviewed-on: http://review.couchbase.org/104751
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>