MB-30029: jemalloc: Restore Windows build steps for v4.1.0
Restore the cbdeps build steps for Windows (using the steps prior to
the 5.0.1 upgrade).

The eventual aim is to upgrade jemalloc for Windows to the same
version as the other platforms (5.1), this patch is a stepping stone
to that goal; to add back build steps which work for the
currently-used version of jemalloc on Windows (4.1.0).

Note this patch creates a -cb3 build - one higher than the current
windows version of -cb2. It /should/ be the same but given it's been
rebuilt we will incrementally update to cb3 to ensure build steps are
correct.

Note(2) this still requires autogen et al from cygwin to build.

Change-Id: I7c81e478264f2cdd026d761aba9f66881afd1496
Reviewed-on: http://review.couchbase.org/105269
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>