CBD-2894: Add support for Miniconda-based Python scripts; deprecate PyInstaller
We build a local minimal Anaconda installer based on Python 3.6.8 and
including the few third-party modules that our code requires. This will
be installed on customer machines as part of the installer.

Deprecate the PyInstall() and PyVenv() functions. Add PyWrapper() function
to create a $PREFIX/bin wrapper script to invoke a given Python 3 script
using the installed Anaconda python.

Change-Id: I96c6793e416d52739676000a3129797428343029
Reviewed-on: http://review.couchbase.org/105693
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>