MB-30040: Build folly package using branch v2019.03.18-couchbase
Update the package as we require the TSan annotations for
SharedMutex.

We now need to explicitly tell folly to compile against C++14 by setting
the CXX_STD flag.

Somebody working on folly decided to add a template parameter to
SharedMutex that defaults to whether or not folly was compiled with or
without TSan. This is a pain for us because we will need different
symbols based on whether or not we are compiling with TSan. Solve this
problem by building and shipping a 'libfollytsan.a' compiled with an
additional '-fsanitize=thread' flag that we can link instead on the
platforms where we will run TSan.

Change-Id: I75a34db07f2c361d9305bc548b7eaaa3382dedab
Reviewed-on: http://review.couchbase.org/107013
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>