MB-30040: Bump boost package revision
We need to link libboost_thread (and libboost_system) for Windows
if we want to use certain folly features. The current boost
package was actually compiled using MSVC2015 (vc140 in lib name).
We need to build the package properly using MSVC2017 to generate
vc141 libs which MSVC2017 will look for when we try to link boost.

Change-Id: I4da2db65206a25cd0f7e4fd3503be0a43144552c
Reviewed-on: http://review.couchbase.org/107450
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>