CBD-3692: Sanitizer fixes for Ubuntu 18.04
This patch contains a backport of the following patches from master to
allow the various Santizers which run as part of commit-validation to
work correctly on Ubuntu 18.04 - which CV has recently been upgraded
to:

1. MB-33716: ThreadSanitizer: Workaround dlopen+RPATH issue (http://review.couchbase.org/107469)
2. MB-32016: Install sanitizer runtime libraries to <prefix>/lib (http://review.couchbase.org/101849)
3. MB-33716: Add dlopen+RPATH workaround for ASan (http://review.couchbase.org/109513)
4. Only create error message if sanitiser library is not found Linux (http://review.couchbase.org/103322)
5. MB-33716: Check '-Wl,--disable-new-dtags' is valid linker flag (http://review.couchbase.org/110040)
6. TSan: Delete RUNPATH from libtsan.so if present (http://review.couchbase.org/112865)
7. ASan / UBSan: Delete RUNPATH from lib<san> if present (http://review.couchbase.org/113137)

The crux of the issue with Ubuntu 18.04 is that the sanitizer runtime
libraries it ships (libtsan.so) contain dynamic linker directives
(RUNPATH) which break how KV-Engine loads engines (e.g. ep.so) and
testsuites. To address this we need to:

a) Use the older RPATH directive in our own code.
b) Install a copy of the sanitizer runtime libs and modify it to
   remove the problematic RUNPATH directive.

Change-Id: Id6c0e71a71b578d674f7f9f777c6d954e7d52a32
Reviewed-on: http://review.couchbase.org/c/tlm/+/138987
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>