Add CMake function add_compile_options_disable_optimization()
Add a new CMake variable - CB_CXX_FLAGS_OPTIMIZE_FOR_DEBUG - which is
defined as the C++ optimization flags to use for targets which it is
unnecessary to perform full optimization on - for example unit tests
or other non-performance sensitive targets.

Use this to define a helper function -
add_compile_options_disable_optimization() which can be used to
disable optimzations for the given directories (and subdirectories).

Change-Id: Id0ee6e7c2004acdfe641edb44fd7270e90defb73
Reviewed-on: http://review.couchbase.org/109029
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>