MB-34015: Upgrade to Python 3.7 for bundled python interpreter
Introduce a .version file in the local cbpy install directory, along with
dependency rules, to ensure that when developers sync a future change to
cbpy, it will both regenerate the cbpy-installer AND re-install it locally.

In addition to updating tlm/python/construct.yaml to declare Python 3.7 and
the related dependent package changes, this required several other changes
for various platform reasons:

 - Windows: This required adding lib\python\runtime\Library\bin to PATH in
   the exe wrapper because Conda changed the rules about where libraries are
   located and how they're loaded with Python 3.7.

 - MacOS: Added a Couchbase-specific build number to the directory used for
   user-specific cbpy installation, to ensure that the new version will be
   installed even if a developer has an earlier "Python for Couchbase Server
   6.5.0" installed.

The latter change in particular pushed me to refactor the CMake code, so
that all logic is now in tlm/python/* rather than split between there and
tlm/cmake/Modules/FindCouchbasePython.cmake. This is clearer anyway because
we don't actually want to "find" Python as part of the build; we want to
define targets and functions for using our own installer.

Change-Id: I84f7b64a554fbfb142c62f6518a6934ab86566cd
Reviewed-on: http://review.couchbase.org/109160
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Kim Trang <kim.trang@couchbase.com>