Debug speedup [1/3] Introduce 'DebugOptimized' build type
+Summary+

First in a series of patches to speed up the regular 'Debug' build
type (as expected to be used in a developers' compile-edit-debug
cycle) by removing -Og from it.

+Background+

Currently the Debug build type enabled _some_ optimization using the
-Og flag where available (see c558d3b).

The primary motivation for the additional of -Og is automated
commit-validation jobs:

Given commit-validation jobs compile once and run once, we want to be
able to balanace compile time and execution time of the unit tests.

However, neither the standard Debug nor Release optimization levels are
optimal :-

- Debug minimises the compile time, but the test runtime (for larger,
  more complex tests) can be /much/ slower than a Release build;
  meaning the overall CV runtime is dominated by test runtime.

- RelWIthDebInfo, while producing fast code at runtime, is slow to
  compile wit production-level optimization, either simply due to its
  size, or things like GMock / GTest template instantiation - the CV
  time is dominated by compile time.

By using a small amount of optimization (e.g -Og) we can achieve a
middle-ground for CV. This is currently implememented by changing the
compilation flags for Debug to use -Og for GCC and Clang.

+Problem+

It turns out that the -Og build variant still adds a non-trivial
amount of compiler runtime compared to a completely unoptimized build
(-O0). For example on macOS 10.13.6 with Apple clang-1000.11.45.5, the
times to build all of kv_engine with -O0 vs -Og are (two runs):

    Debug Optimized (-Og)

      351.51 real      2922.75 user       162.72 sys
      350.67 real      2930.46 user       160.42 sys

    Debug (-O0)

      270.03 real      2221.26 user       153.19 sys
      269.23 real      2216.17 user       152.69 sys

Debug is 1.3x faster than Debug Optimized (both real & user+sys).

(The difference is even more pronounced in large test files - for
example vbucket_durability_test.cc is 1.7x faster to build as -O0 vs
-Og:

    Debug Optimized single source file:

      9.62 real         9.27 user         0.31 sys
     10.72 real         9.12 user         1.28 sys

    Debug single source file:

      5.60 real         5.21 user         0.30 sys
      5.87 real         5.25 user         0.41 sys

As such, the current Debug==-Og can be undesirable from a developer's
workflow; however we _do_ still want the ability to build in CV with
-Og.

+Solution+

This patch introduces a new build type - DebugOptimized - which is
basically what Debug currently is - it compiles with -Og.

This allows use-cases which do want a balance between compile-time and
execution speed to explcitly request that.

Change-Id: I5de742b38dd8510852856440261e8df1b5928288
Reviewed-on: http://review.couchbase.org/113127
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>