MB-35492: TSan: Pass TSAN_FLAG to Golang CGO code
If ThreadSanitizer is enabled, then Golang code which uses CGO to
access C libraries (e.g. sigar) can be flagged by TSan as containing
data races even when there are no problems in the usage.

This is due to the TSAN compiler flag (-fsanitize=thread) _not_
getting passed down to Golang/CGO; as a result the synchronisation in
the Go code / runtime is not visible to TSan, and races are reported.

Fix by passing THREAD_SANITIZER_FLAG down to the Go compiler via
CGO_CFLAGS env var.

Change-Id: Iec820af155fa97345653a8fc916fef0c8a9fffff
Reviewed-on: http://review.couchbase.org/113422
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Lilei Chen <lilei.chen@couchbase.com>