MB-37096: Build debug variant of Boost on Windows
In addition to the current Release variant of Boost, also build the
debug variant on Windows.

The debug variant is linked against the MSVC Debug CRT, which enables
various extra assertions in the C runtime, and crucially allows us to
link against Boost when the _application_ (e.g. memcached) is linked
with Debug CRT.

Change-Id: Idc2cf19e39ea30133fcba796302bebdccd3532e0
Reviewed-on: http://review.couchbase.org/119424
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>