MB-37096: Build debug variant of flatbuffers on Windows
In addition to the current Release variant of flatbuffers, also
build the debug variant on Windows.

The debug variant is linked against the MSVC Debug CRT, which allows
us to link against flatbuffers when the application (e.g. memcached) is
linked with Debug CRT.

Change-Id: Id8e8110e19d59a092b28441eab3ba3f7b3d18a55
Reviewed-on: http://review.couchbase.org/119429
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>