MB-37096: Teach CMake to find Debug variant of jemalloc
Update FindCouchbaseJemalloc.cmake to be able to locate a Debug build
variant of jemalloc.

JEMALLOC_LIBRARY_RELEASE and JEMALLOC_LIBRARY_DEBUG are added to refer
to release and debug jemalloc libraries respectively (if present).

JEMALLOC_LIBRARIES is now set such that the correct variant
(debug/release) is used based on CMAKE_BUILD_TYPE.

Note that currently no debug build of jemalloc exists in cbdeps, so
JEMALLOC_LIBRARY_DEBUG will not be set and JEMALLOC_LIBRARIES will
always refer to the release build; however in subsequent patches this
will change (and the build location of Release variant will also change)
so this patch needs to be in place beforehand.

Change-Id: I8fc1535fefa65a49271ce4d6751c0f704b18ef17
Reviewed-on: http://review.couchbase.org/119421
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>