MB-37096: Use folly v2019.08.12.00-cb2 & glog 0.3.5-cb4
Update to folly v2019.08.12.00, cb2. This includes the debug build of
folly in addition to a release variant.

Also upgrade glog to 0.3.5-cb4 (dependancy of folly).

This requires changes to the FindCouchbaseFolly.cmake script, to search
for both the release and debug versions of the library.

Change-Id: Ibe30a87de4a0a872abb04dfa5fb215ae40c42569
Reviewed-on: http://review.couchbase.org/119428
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>