MB-37137: Fix rpath for libevent on macosx
There is something wrong in the way libevent is being
built with cmake (with respect to RPATH) as it is
being stripped off as part of make install.

I don't have access to a mac machine for the moment
(and given that CC is broken) I chose the easy way
out to fix this after make install (but before we
package up the binaries).

Change-Id: Ib255c6c194396fbef4929d7244c1e2180799964d
Reviewed-on: http://review.couchbase.org/118858
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>