MB-37956: Support both v0.25 and v0.36 of gsl-lite
gsl-lite has changed it's API between versions 0.25.0 and 0.36.0. To
allow us to upgrade to 0.36.0 without having to make simultaneous
patches to the various porjects which use it, modify our
third-party/CMakeLists.txt which includes it to perform the necessary
different config setup between the two versions.

Once we have upgraded to 0.36 the old code for 0.25 can be removed.

Change-Id: Ibee979fd504ae805e737f2463eefe4da227e8e61
Reviewed-on: http://review.couchbase.org/122419
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>