MB-37898: Remove gflags from folly (and build) [take 2]
[Re-apply after previous issues with Clang-analyser]

*Problem*

As part of the upgrade to C++17, Folly has a link-time dependancy on
GFlags - i.e. any executable linking Folly also must link GFlags.

However, this results in some nasty linking issues due to the fact
that we use a mix of static and shared libraries. The problem
manifests as the following fatal error when attempting to /run/ any
program which has a mix of static and shared libraries
(e.g. memcached_testapp):

    ERROR: something wrong with flag 'flagfile' in file 'gflags.cc'.
    One possibility: file 'gflags.cc' is being linked both statically
    and dynamically into this executable.

As per the hint in the message, the problem is related to how how
GFlags uses static initialization to auto-register flags the developer
has added to their code. In the case of memcached_testapp (and other
binaries), we end up with multiple instances of gflags being linked,
and consequently multiple attempts to register the same flags:

    memcached_testapp
        ^
	libfolly.a [statically linked to libfolly.a]
        libdefault_engine.so
	    ^
            platform.so [statically linked to libfolly.a]

i.e. there's two copies of gflags in the memcached_testapp process at
runtime.

To solve this we need to remove the mix of libaries - either have all
libraries (using folly) statically linked, or all shared.

This isn't particulary desirable however - folly itself doesn't
recommend / support shared linking; but we cannot be fully statically
linked as for example libcouchstore.so must exist to be loaded by
python (at runtime).

In the particlar instance of libdefault_engine.so above we could
change it's linking to static, but I believe similar problems exist
with libep.so etc, and again that is harder to link statically.

*Solution*

Given that gflags is somewhat optional in Folly (it's used in only a
couple of places and it's CMakefiles don't strictly require it),
address this problem by removing gflags from Folly (and consequently
our own build). This completely avoids the whole problem :)

This requires a small modification to Folly (see associated patch
file) to not set GFlags-related variables if not found, and to skip
compiling the couple of source files which _do_ require it.

Change-Id: Ia9dc0cd8cc5344a05cb63f5a04f363e255565308
Reviewed-on: http://review.couchbase.org/123068
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>