MB-37898: Explictly check for required C++17 headers
Some compilers *couch* AppleClang 9 *couch* claim support C++17 but
are missing core library features such as <optional> and <variant>.

This can result in misleading / confusing output when attempting to
build with such a compiler - the configure step passes, but the build
will fail halfway through when the first use of std::optional /
std::variant is encountered.

To aid the user, add an explicit check for the known missing headers
and fail the configure step if not found.

Change-Id: I69d902407461b8159ce6c262ef71812bae27202b
Reviewed-on: http://review.couchbase.org/c/tlm/+/124496
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>