CBD-2895: Improvements in cbdeps .md5 handling
A cbdeps .tgz on S3 always has a corresponding .md5. However cbdeps really
wanted it to be named .tgz.md5, which led to weird inconsistencies. It
turned out to be especially annoying with cbdeps 2.0 toy builds as you
would need to manually rename the downloads from latestbuilds before
putting them in .cbdepscache, so it seemed like a good time to fix it.

Change-Id: Ib59885fe0604b43b84caa2e27ecc16446c61757f
Reviewed-on: http://review.couchbase.org/c/tlm/+/126645
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>