MB-39537: Erlang/OTP "cheshirecat"
Based on internal discussion, we commit to having separate git branches
in the Erlang repository for each Server release, so that we can
continue naming said branches after Server releases. As such, and since
the number of changes is slowly becoming more significant, we are also
naming the cbdeps packages themselves with Server release-name versions.
So this commit introduces version "cheshirecat" of Erlang, which is
currently based on OTP-22.3.4.6.

Change-Id: Id0e8f5d03f7b6e98817fbf8f563661a32c874e24
Reviewed-on: http://review.couchbase.org/c/tlm/+/131633
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>