MB-41209: cbdeps: Update to Folly v2020.03.05.00
This is an intermediate step to upgrading to the latest folly
release. We cannot easily go to the latest (v2020.08.24.00) as that
removes some classes KV-Engine currently uses (CachelinePadded), and
the replacement (cacheline_aligned) doesn't exist yet in our current
version (v2019.08.12.00).

As such, upgrade to this intermediate version which has both the old
(CachelinePadded) and new (cacheline_aligned) classes, so we can
migrate KV-Engine over without having to synchronize watches^Wcheckins
across multiple repos.

Changes:

- Remove the custom patch to make GFlags optional as this version of
  Folly has it included.

- Add a patch to fix compilation error on CentOS7+GCC7.3 due to that
  environment needing __STDC_FORMAT_MACROS defined for PRId64 to be
  present.

- Make Boost path explicit to ensure any system Boost is not
  incorrectly picked up.

- Add location of fmt's fmt-config.cmake to CMAKE_PREFIX_PATH.

- Fix Windows openssl dependancy to use same version as toplevel
  manifest.cmake.

- Remove CXX_STD=c++17 as Folly now builds as C++17 by default.

Change-Id: Iac19fec572275cc6bbe7b5693250174e89c9b1f6
Reviewed-on: http://review.couchbase.org/c/tlm/+/135228
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>