MB-41209: Define WIN32_LEAN_AND_MEAN under MSVC
When compiling with Folly v2020.03.05.00 or newer on Windows, compiler
errors are seen a Folly header which includes glog/logging.h is used:

    build\tlm\deps\glog.exploded\include\glog/log_severity.h(51):
    fatal error C1189:
    #error:  ERROR macro is defined. Define GLOG_NO_ABBREVIATED_SEVERITIES before including logging.h. See the document for detail.

This is doucmented in the google log docs[1], essentially Windows.h
defines a macro which collides with one defined by glog.

There's two ways to fix this:

1. Disable glog's abbeviated severity levels by defining
   GLOG_NO_ABBREVIATED_SEVERITIES.

2. Disable the definition of ERROR - which can be done by defining
   WIN32_LEAN_AND_MEAN.

(2) has the advantage of also cutting a load of Windows.h content
which we don't use (for example various graphical things). As such,
solve via (2).

[1]: http://rpg.ifi.uzh.ch/docs/glog.html

Change-Id: I01338d47b2943e3646e959f19bf6118428ce0e4d
Reviewed-on: http://review.couchbase.org/c/tlm/+/135468
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>