MB-40090: Update to OpenSSL 1.1.1h
Since Erlang on Windows links OpenSSL statically, we also update to a
new build number for Erlang. Did so on all platforms to keep things
consistent, even though on Linux/Mac it's not strictly necessary.

Improved Erlang Windows build to dynamically install the right version
of OpenSSL rather than expecting it to exist on the build VM.

Updated mentions of OpenSSL in folly and libevent build scripts, but
didn't actually rebuild since OpenSSL is not actually included in
those packages.

Change-Id: Ic2f6b3c285c926d4521ddabc417598716cb9235c
Reviewed-on: http://review.couchbase.org/c/tlm/+/136952
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>