MB-41896: Link to shared TSan runtime under Clang
Clang links to the TSan runtime library statically by default. This is
problematic as we can have multiple different libraries linked into a
final executable, and hence end up with multiple copies of the static
runtime library.  This causes incorrect behavour when TSan runtime
tries to track locks. Specifically, it results in TSan internal checks
failing in random places when locks are acquired - for example:

    FATAL: ThreadSanitizer CHECK failed: compiler-rt/lib/sanitizer_common/sanitizer_deadlock_detector1.cc:166 "((!dd.isHeld(&lt->dd, m->id))) != (0)" (0x0, 0x0)

Link to the shared runtime library to avoid this problem.

Change-Id: I7c7a7bd53ac67dee82c20eed238213607ba53e41
Reviewed-on: http://review.couchbase.org/c/tlm/+/137766
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>