MB-41988: Locate sanitizer runtime libs via dynamic linker
The tlm CMake function install_sanitizer_library() locates the wrong
version of the TSan runtime library when building with Clang - instead
of finding the GCC 7.3 library, it finds the original Ubuntu 16.04 GCC
5 one:

    -- Found TSan at: /usr/lib/x86_64-linux-gnu/libtsan.so.0 installing to: /home/couchbase/couchbase/install/lib

However the newer GCC 7.3.0 library is located under /usr/local/lib64.

This appears to be an issue in how find_file() handles relative paths
in conjunction with symbolic links - it doesn't correctly resolve
/usr/lib/gcc/x86_64-linux-gnu/7.3.0/../../../../lib64 which _should_
resolve to /usr/local/lib64.

However, the current method is bittle in general - there's no
guarantee that the shared library the runtime linker will find is the
same as the compile-time linker.

To address this, change the method used to locate the
sanitizer. Compile and link a test program with the appropriate
sanitizer, then check the output of the dynamic linker to determine
the absolute file path.

Change-Id: I0bf8442fc40bacf5e8965af61213f2d7575b4d78
Reviewed-on: http://review.couchbase.org/c/tlm/+/137964
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>