CMake: find unsanitized Folly library for ThreadSanitizer
When compiling with ThreadSanitizer, parts of platform are built
'unsanitized', so they can be correctly loaded into executables which
are not compiled with TSan support (e.g. libcouchstore.so loaded into
beam.smp for the view engine).

However, on macOS we see runtime linker errors for symbols in folly
when linking platform_unsanitized:

    Undefined symbols for architecture x86_64:
      "folly::CacheLocality const& folly::CacheLocality::system<std::__1::atomic>()", referenced from:
          cb::get_num_last_level_cache() in sysinfo.cc.o
      "folly::AccessSpreader<std::__1::atomic>::getcpuFunc", referenced from:
          cb::stripe_for_current_cpu(unsigned long) in sysinfo.cc.o

    (It's not clear why such errors are not seen for Linux...)

This is due to platform_unsanitized not being linked against folly -
note it cannot use the existing FOLLY_LIBRARIES variable as that is
compiled with TSan which cannot be linked into the
*unsanitized_libraries.

As such, define a new FOLLY_LIBRARIES_UNSANITIZED variable which
contains the non-TSan folly library, for platform to then link
against.

Change-Id: I707e1de0fd02100ae188216682caf5c91f7906d9
Reviewed-on: http://review.couchbase.org/c/tlm/+/137998
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>