cbdeps Folly: use 'couchbase' branch for patches
Folly requires a couple of small patches to the source to address
issues in how we build it. These were originally added as .patch files
on the assumption they would be upstreamed quickly. However, they
still haven't been accepted upstream and we are seeing problems
applying the patch files on Windows.

As such, create our own 'v2020.08.24.00-couchbase' branch in
cbdeps/folly and apply the two patches needed there, removing the need
for the patch files.

Change-Id: Idb2f6fb77837c91538d49a1aebffcc2455ba9cf4
Reviewed-on: http://review.couchbase.org/c/tlm/+/138812
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>