GCC: Default to non-PIE
If building with a version of GCC which defaults to PIE code
(--enable-default-pie), we can encounter linker errors when linking
against precompiled static libraries similar to:

    /usr/bin/ld: foo.a(bar.cc.o): relocation R_X86_64_32 against `.rodata.str1.8' can not be used when making a PIE object; recompile with -fPIC

For example, the GCC which ships with Ubuntu 18.04 suffers from this
issue.

Address this by disabling the effect of '--enable-default-pie' - set
the default to non- position-independent executables.

Change-Id: Ia3d057dc77f00d5775ad2ca9c56b21f52429fa5f
Reviewed-on: http://review.couchbase.org/c/tlm/+/140650
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>