MB-42977: cbdeps: Upgrade to Boost 1.74 [1/2]
boost::process (as required by cluster_testapp) requires headers from
asio and fusion dirs, so need to rebuild boost to not remove them from
the headers which are archived.

However, we _also_ need to upgrade to Boost 1.74; as the previous
version (1.67) doesn't build with XCode 11.3.1 as used by the macOS
cbdeps builder.

This is a multi-step process:

1. Build Boost 1.74; make available as cbdep (this patch)

2. Rebuild Folly which depends on Boost, make folly available as cbdep
   and use new versions of Boost and Folly.

Change-Id: I19dbd9dabee41746c08025f2e51d86ede80d73d9
Reviewed-on: http://review.couchbase.org/c/tlm/+/141153
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>