MB-45057: Update to latest Python packages
This does NOT currently update to a newer version of Python itself yet
(which will also require updating some dependent packages). It brings
in changes that I made and then rolled back for 6.6.2, namely updating
to the "noarch" request 2.25.1 along with the CMake changes necessary
for that to work. It also updates pip which had shown a security
vulnerability.

Change-Id: Ib13384dc8a439f7b35ff93b52ebbeb3dbe2291aa
Reviewed-on: http://review.couchbase.org/c/tlm/+/148926
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>