MB-45440, CBD-3755: Convert cbpy to cbdeps; exclude chardet, readline, pip
To avoid GPL/LGPL concerns, build stub versions of the conda "chardet"
and "readline" packages, as well as "pip" since the conda pip package
includes a vendored copy of chardet. Server scripts do not depend on any
of those packages.

Because this makes building cbpy considerably more complex, make cbpy a
cbdeps package to split it out of the Server build process. I used
cbdeps 1.0 here so that the corresponding Black Duck manifest could be
kept here in tlm right alongside cbpy's package manifest, so it will be
properly versioned.

Change-Id: Ibc97c30e0b3b56a0ec4faa0d5fca2e09bf8b65a2
Reviewed-on: http://review.couchbase.org/c/tlm/+/150292
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Blair Watt <blair.watt@couchbase.com>