MB-45378: Cleanup Folly include settings
Previously the FOLLY_INCLUDE_DIR variable was used to set include
paths for "modern CMake" targets. This had issues with repeated runs
of CMake, as the variable's value is modified after it's initial uses
- where on the first run a dependancy include path wasn't associated
with Folly::headers but would be on later runs.

Use the Folly::headers target to 'own' the includes needed for Folly,
then have the other folly targets depend on that; avoiding entirely
FOLLY_INCLUDE_DIR.

See also the linked patches to couchstore, platform and kv_engine with
same Change-Id.

Change-Id: I9e0f9474e93ceddc92d324406308688b545e5c4a
Reviewed-on: http://review.couchbase.org/c/tlm/+/151242
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>