MB-46988: Fix PCH headers causing ccache cache misses
When precompiled headers are enabled (CB_PCH=ON), the cache hit rate
of CCache drops to less tha 100%, even if no source changes have been
made between builds.

Instead, 'ccache -s' reports 'preprocessor error':

    cache hit (direct)                     0
    cache hit (preprocessed)               0
    cache miss                             1
    cache hit rate                      0.00 %
    preprocessor error                     1

Enabling ccache debug mode and examinind a log file for a failing
object, we see the following:

    [2021-06-17T10:57:43.280809 15720] Running preprocessor
    [2021-06-17T10:57:43.280850 15720] Executing /usr/bin/c++ <cut> -E ../platform/src/je_arena_malloc_stats.cc
    [2021-06-17T10:57:43.508729 15720] Preprocessor gave exit status 1
    [2021-06-17T10:57:43.508839 15720] Failed; falling back to running the real compiler
    [2021-06-17T10:57:43.509102 15720] Result: preprocessor error

i.e. it failed to invoke the preprocessor (via c++ -E) and hence
couldn't lookup the file in the cache.

Manually running the (omitted above) full command-line of c++ shows
the reason why the preprocessor failed:

    In file included from <built-in>:2:
    In file included from /Users/dave/cb_server_S1_testing/build/platform/CMakeFiles/platform_pch_fpic.dir/cmake_pch.hxx:5:
    In file included from /Users/dave/cb_server_S1_testing/platform/precompiled_headers.h:14:
    tlm/deps/folly.exploded/include/folly/portability/GTest.h:32:10: fatal error: 'gtest/gtest.h' file not found
    #include <gtest/gtest.h>
             ^~~~~~~~~~~~~~~
    1 error generated.

Indeed, examining the include paths passed to the original c++
invocation, there is no include path for googletest.

What seems to be happening here is when the compiler attempts to
compile a source file (e.g. je_arena_malloc_stats.cc) and PCH is
enabled, the precompiled header is passed to it and can be consumed
without needing access to the original header source. This seems
reasonble; given the whole point of the PCH is to pre-compile the
original header files. However, when ccache attempts to invoke the
preprocessor via -E; the compiler _does_ need to locate the original
header source.

To fix this issue, explicilty add the neccesary header include paths
to all targets depending _on_ the PCH. This ensurs that the header
path is present, and ccache succeeds in invoking the preprocessor:

    [2021-06-17T13:21:49.941158 28588] Running preprocessor
    [2021-06-17T13:21:49.941173 28588] Executing /usr/bin/c++ ... \
        -Xclang -include -Xclang build/platform/CMakeFiles/platform_pch_fpic.dir/cmake_pch.hxx \
        <cut> \
        -I../third_party/googletest/googletest/include -Xclang -include-pch \
        -Xclang build/platform/CMakeFiles/platform_pch_fpic.dir/cmake_pch.hxx.pch \
        -fpch-preprocess -E ../platform/src/je_arena_malloc_stats.cc
    [2021-06-17T13:21:50.206499 28588] Include file /Users/dave/cb_server_S1_testing/build/platform/CMakeFiles/platform_pch_fpic.dir/cmake_pch.hxx.pch too new
    [2021-06-17T13:21:50.206562 28588] Got result name from preprocessor

... and this results in 100% cache hit rate (assuming no source has
changed):

    $ ninja platform/install && ccache -z && ninja -t clean platform/install && ninja platform/install && ccache -s
    cache directory                     /Users/dave/.ccache
    primary config                      /Users/dave/.ccache/ccache.conf
    secondary config (readonly)         /usr/local/Cellar/ccache/4.3/etc/ccache.conf
    stats updated                       Thu Jun 17 13:50:47 2021
    stats zeroed                        Thu Jun 17 13:50:45 2021
    cache hit (direct)                    95
    cache hit (preprocessed)               0
    cache miss                             0
    cache hit rate                    100.00 %
                                      ^^^^^^
    called for link                       34
    called for preprocessing              16

Change-Id: Ib567be1354f9ccd98ebbbd31db40f9deb77a2eb3
Reviewed-on: http://review.couchbase.org/c/tlm/+/156091
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>